requirements.txt
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.10.0
neo4j>=5.8.0
flask>=2.0.0